df: Fix refactoring error; Roll back some changes from export branch
