+++ Cargo.toml
@@ -9,5 +9,4 @@ members = [
   "libDF",
   "pyDF",
   "pyDF-data",
-  "df-tract"
 ]
