+++ libDF/src/lib.rs
@@ -18,6 +18,7 @@ pub mod transforms;
 mod reexport_dataset_modules {
     pub mod augmentations;
     pub mod dataset;
+    pub mod mix_utils;
     pub mod util;
     pub mod wav_utils;
 }
